Validations added

+++ src/main/java/com/example/longest_playing_pairs/repository/PlayerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface PlayerRepository extends JpaRepository<Player, Integer> {
+public interface PlayerRepository extends JpaRepository<Player, Long> {
     Optional<Player> findById(Long id);
 }
